CLEANUP: Refactored SMGetFuture.
